+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 **This version contains BREAKING CHANGES if you are using them.**
 
-- [world] refactor: make `per-dimension` gamerules as a standalone `world.gamerule` module.
+- [world] refactor: make `per-dimension gamerules` as a standalone `world.gamerule` module.
   - This simplifies the configs of `world` module.
   - It allows you to define `per-dimension gamerules` for `any dimension`, even the `vanilla dimensions`.
   - It allows you to use the `per-dimension gamerules` with other `world management mod`.